-semi-fixed crouching, now properly changes physics shape (character's physicsl height cuts in half)     -however, graphical glitches still exist -removed an extra physics_engine.step from handle_key_press() in main, was causing player to move twice as fast as other objects -adjusted constants related to player movement
